requirements.txt
asgiref==3.7.2
attrs==23.1.0
bcrypt==4.0.1
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
dj-database-url==2.0.0
Django==4.2.3
django-betterforms==2.0.0
django-crum==0.7.9
django-multiform==0.1
django-rest-framework==0.1.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
isodate==0.6.1
lxml==4.9.3
mssql-django==1.3
packaging==23.1
paramiko==3.3.1
Pillow==10.0.0
platformdirs==3.10.0
psycopg2-binary==2.9.7
pycparser==2.21
PyNaCl==1.5.0
pyodbc==4.0.39
pytz==2023.3
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
zeep==4.2.1
